pyproject.toml
[tool.poetry]
name = "object-detection-system"
version = "0.1.0"
description = ""
authors = ["Najib Aghenda <najib.aghenda@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.1"
torchvision = "^0.18.1"
streamlit = "^1.36.0"
opencv-python = "^4.10.0"
pillow = "^10.4.0"
numpy = "^2.0.0"
requests = "^2.32.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
